made some bunch of corrections after moving siome code to Effort.Next

+++ src/CExetnsions.Test.Model/Author.cs
@@ -8,7 +8,7 @@ namespace CExtensions.Test.Model
 {
     public partial class Author
     {
-        public decimal Id { get; internal set; } // AUT_ID (Primary key)
+        public Int64 Id { get; internal set; } // AUT_ID (Primary key)
         public string FirstName { get; set; } // AUT_FIRSTNAME
         public string LastName { get; set; } // AUT_LASTNAME
         public decimal? Experience { get; set; } // AUT_EXPERIENCE
