+++ src/CExetnsions.Test.Model/Post.cs
@@ -8,12 +8,12 @@ namespace CExtensions.Test.Model
 {
     public class Post
     {
-        public decimal Id { get; set; } // Post_Id (Primary key)
+        public Int64 Id { get; set; } // Post_Id (Primary key)
         public string Subject { get; set; } // Subject
         public string Body { get; set; } // Body
         public DateTime? DateCreated { get; set; } // DateCreated
         public DateTime? DateModified { get; set; } // DateModified
-        public decimal AutId { get; set; } // AUT_ID
+        public Int64 AutId { get; set; } // AUT_ID
 
         // Foreign keys
         public virtual Author Author { get; set; } // FK_POST_AUTHOR
