+++ src/CExtensions.Sample.Model/Author.cs
@@ -8,7 +8,7 @@ namespace CExtensions.Sample.Model
 {
     public class Author
     {
-        public decimal Id { get; internal set; } // AUT_ID (Primary key)
+        public Int64 Id { get; internal set; } // AUT_ID (Primary key)
         public string FirstName { get; set; } // AUT_FIRSTNAME
         public string LastName { get; set; } // AUT_LASTNAME
         public decimal? Experience { get; set; } // AUT_EXPERIENCE
